Game redesigned using OOP

+++ makefile
@@ -1,19 +1,22 @@
-optFlags = -Ofast -std=c++11 -Wall
-extraFlags = -c -Wall
+optFlags = -Ofast -std=c++11 -w
+extraFlags = -c -w
 oglFlags = $(optFlags) -lglut -lGL -lGLU -lm
 includePath = ./source/include
 
-Simulator : Object3D.o Parser.o
-	g++ ./source/main.cpp $(oglFlags) ./bin/*.o -o simulator.out
+Simulator : Game.o
+	g++ ./source/main.cpp ./bin/*.o $(oglFlags) -o simulator.out
+
+Game.o : Parser.o Object3D.o
+	g++ $(includePath)/Game.cpp $(extraFlags) -o ./bin/Game.o
+
+Parser.o : Object3D.o
+	g++ $(includePath)/Parser.cpp $(extraFlags) -o ./bin/Parser.o
 
 Object3D.o :
 	g++ $(includePath)/Object3D.cpp $(extraFlags) -o ./bin/Object3D.o
-
-Parser.o :
-	g++ $(includePath)/Parser.cpp $(extraFlags) -o ./bin/Parser.o
 
 start :
 	./simulator.out
 
 clean :
-	rm ./bin/*.o *.out
+	clear && rm ./bin/*.o *.out
